requirements.am.txt
asn1crypto==0.24.0
cryptography==2.7
PyJWT==1.7.1
argon2-cffi==19.1.0
cffi==1.12.3
pycparser==2.19
passlib==1.7.1
urllib3==1.25.6
falcon==2.0.0
pymongo==3.9.0
minio==5.0.1
gunicorn==19.9.0
pytz==2019.2
tzlocal==2.0.0